reuse sample data in 2 tests
